Update Main Menu
First, we _install_ ClassicPress.
Then, we _use_ it.
Then we develop with it.
Perhaps, we want to test it.
We may have Faqs.

I think all the rest should not be on this page.
Maybe, we may want to Develop _with_ ClassicPress, thus, add a link to a subpage where:
- developer documentation (code generated Codex) is linked.
- more advanced developer oriented docs such as Security Page, etc.

The contents of _Developing ClassicPress_ should be in the "Contributors" menu page > https://www.classicpress.net/contributors/ or in https://www.classicpress.net/community/ and likely not be linked here. 
It is not telling us "how to". 
It just tells us that you may help and that changes go thru a poll/vote. But it wont help neither a user or a developer to actually achieve something with CP.

Brand Guidelines should, again, not be in a User Doc. They are already in the main site and that is enough.